roles

+++ fund-me/routes/web.php
@@ -21,22 +21,22 @@ Route::get('/', function () {
 
 Route::prefix('stories')->name('stories-')->group(function () {
     Route::get('/', [S::class, 'index'])->name('index');
-    Route::get('/create', [S::class, 'create'])->name('create');
-    Route::post('/create', [S::class, 'store'])->name('store');
+    Route::get('/create', [S::class, 'create'])->name('create')->middleware('role:admin|client');
+    Route::post('/create', [S::class, 'store'])->name('store')->middleware('role:admin|client');
     Route::get('/{story}', [S::class, 'show'])->name('show');
-    Route::get('/edit/{story}', [S::class, 'edit'])->name('edit');
-    Route::put('/edit/{story}', [S::class, 'update'])->name('update');
-    Route::delete('/delete/{story}', [S::class, 'destroy'])->name('delete');
+    Route::get('/edit/{story}', [S::class, 'edit'])->name('edit')->middleware('role:admin');
+    Route::put('/edit/{story}', [S::class, 'update'])->name('update')->middleware('role:admin');
+    Route::delete('/delete/{story}', [S::class, 'destroy'])->name('delete')->middleware('role:admin');
     });
 
 Route::prefix('hashtags')->name('hashtags-')->group(function () {
-    Route::get('/', [H::class, 'index'])->name('index');
-    Route::get('/create', [H::class, 'create'])->name('create');
-    Route::post('/create', [H::class, 'store'])->name('store');
-    Route::get('/{hashtag}', [H::class, 'show'])->name('show');
-    Route::get('/edit/{hashtag}', [H::class, 'edit'])->name('edit');
-    Route::put('/edit/{hashtag}', [H::class, 'update'])->name('update');
-    Route::delete('/delete/{hashtag}', [H::class, 'destroy'])->name('delete');
+    Route::get('/', [H::class, 'index'])->name('index')->middleware('role:admin');
+    Route::get('/create', [H::class, 'create'])->name('create')->middleware('role:admin');
+    Route::post('/create', [H::class, 'store'])->name('store')->middleware('role:admin');
+    Route::get('/{hashtag}', [H::class, 'show'])->name('show')->middleware('role:admin');
+    Route::get('/edit/{hashtag}', [H::class, 'edit'])->name('edit')->middleware('role:admin');
+    Route::put('/edit/{hashtag}', [H::class, 'update'])->name('update')->middleware('role:admin');
+    Route::delete('/delete/{hashtag}', [H::class, 'destroy'])->name('delete')->middleware('role:admin');
     });
 
 Auth::routes();
